requirements.txt
streamlit
langchain
unstructured
faiss-cpu
sentence-transformers
langchain_groq
python-dotenv
langchain_community